packages/try-nats/py/pyproject.toml
[project]
name = "try-nats"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Henry.Huang", email = "hhglory@outlook.com"},
]
dependencies = [
    "nats-py>=2.8.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.4.2",
    "mypy>=1.10.1",
    "isort>=5.13.2",
    "pytest>=8.2.2",
]
